requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
discord.py==1.5.1
idna==2.10
multidict==4.7.6
praw==7.1.0
prawcore==1.5.0
python-dotenv==0.15.0
requests==2.25.1
six==1.15.0
update-checker==0.18.0
urllib3==1.26.2
websocket-client==0.57.0
yarl==1.5.1
